+++ examples/custom_readers.rs
@@ -1,11 +1,6 @@
 use std::path::PathBuf;
 
-use rpkl::{
-    api::external_reader::{
-        outgoing::PathElements, reader::ExternalReaderRuntime, PklModuleReader, PklResourceReader,
-    },
-    EvaluatorOptions,
-};
+use rpkl::{api::reader::*, EvaluatorOptions};
 use serde::Deserialize;
 
 pub struct LdapReader;
@@ -36,7 +31,7 @@ impl PklResourceReader for LdapsReader {
         Ok(uri.bytes().collect())
     }
 
-    fn list(&self, uri: &str) -> Result<Vec<PathElements>, Box<dyn std::error::Error>> {
+    fn list(&self, _uri: &str) -> Result<Vec<PathElements>, Box<dyn std::error::Error>> {
         Ok(vec![])
     }
 }
@@ -46,7 +41,7 @@ impl PklModuleReader for ModuleReader {
         "remote"
     }
 
-    fn read(&self, uri: &str) -> Result<String, Box<dyn std::error::Error>> {
+    fn read(&self, _uri: &str) -> Result<String, Box<dyn std::error::Error>> {
         Ok("".to_string())
     }
 
@@ -60,6 +55,7 @@ impl PklModuleReader for ModuleReader {
 }
 
 pub fn main() {
+    #[allow(dead_code)]
     #[derive(Debug, Deserialize)]
     struct Config {
         username: String,
@@ -73,7 +69,6 @@ pub fn main() {
         .join("external-reader.pkl");
 
     let options = EvaluatorOptions::default()
-        .properties([("name", "Ferris")])
         .add_client_module_readers(ModuleReader)
         .add_client_resource_readers((LdapReader, LdapsReader));
     let config: Config = rpkl::from_config_with_options(path, options).unwrap();
