+++ src/api/evaluator.rs
@@ -10,63 +10,44 @@ use crate::{
         handle_list_modules, handle_list_resources, handle_read_module, handle_read_resource,
     },
     error::{Error, Result},
+    internal::msgapi::{
+        incoming::PklServerMessage,
+        outgoing::{CloseEvaluator, CreateEvaluator, EvaluateRequest, ExternalReader},
+        PklMessage,
+    },
     utils::{self, macros::_debug},
 };
 
-use outgoing::{
-    codes::{
-        LIST_MODULES_REQUEST, LIST_RESOURCES_REQUEST, READ_MODULE_REQUEST, READ_RESOURCE_REQUEST,
-    },
-    CloseEvaluator, CreateEvaluator, EvaluateRequest,
+use crate::internal::msgapi::codes::{
+    LIST_MODULES_REQUEST, LIST_RESOURCES_REQUEST, READ_MODULE_REQUEST, READ_RESOURCE_REQUEST,
 };
-use responses::PklServerMessage;
-use serde::{Deserialize, Serialize};
 
 use crate::{api::decoder::pkl_eval_module, pkl::PklMod};
-
-pub mod outgoing;
-pub mod responses;
 
 #[cfg(feature = "trace")]
 use tracing::debug;
 
-use super::{
-    external_reader::{
-        reader::{IntoModuleReaders, IntoResourceReaders},
-        PklModuleReader, PklResourceReader,
-    },
-    msgapi::PklMessage,
-};
+use super::reader::{IntoModuleReaders, IntoResourceReaders, PklModuleReader, PklResourceReader};
 
 pub(crate) const EVALUATE_RESPONSE: u64 = 0x24;
 pub(crate) const CREATE_EVALUATOR_REQUEST_ID: u64 = 135;
 pub(crate) const OUTGOING_MESSAGE_REQUEST_ID: u64 = 9805131;
 
-#[derive(Serialize, Deserialize)]
-pub struct ExternalReader {
-    /// May be specified as an absolute path to an executable
-    /// May also be specified as just an executable name, in which case it will be resolved according to the PATH environment variable
-    pub executable: String,
-
-    /// Command line arguments that will be passed to the reader process
-    pub arguments: Vec<String>,
-}
-
 // options that can be provided to the evaluator, such as properties (-p flag from CLI)
 pub struct EvaluatorOptions {
-    /// Properties to pass to the evaluator. Used to read from `props:` in `.pkl` files.
+    /// Properties to pass to the evaluator. Used to read from `props:` in `.pkl` files
     pub properties: Option<HashMap<String, String>>,
 
-    /// Client-side module readers.
+    /// Client-side module readers
     pub client_module_readers: Option<Vec<Box<dyn PklModuleReader>>>,
 
-    /// Client-side resource readers.
+    /// Client-side resource readers
     pub client_resource_readers: Option<Vec<Box<dyn PklResourceReader>>>,
 
-    /// External resource readers.
+    /// External resource readers
     pub external_resource_readers: Option<HashMap<String, ExternalReader>>,
 
-    /// External module readers.
+    /// External module readers
     pub external_module_readers: Option<HashMap<String, ExternalReader>>,
 }
 
@@ -139,13 +120,21 @@ impl EvaluatorOptions {
     pub fn external_resource_reader(
         mut self,
         key: impl Into<String>,
-        reader: ExternalReader,
+        executable: impl Into<String>,
+        args: &[&str],
     ) -> Self {
+        let reader = ExternalReader {
+            executable: executable.into(),
+            arguments: args
+                .into_iter()
+                .map(|s| s.to_string())
+                .collect::<Vec<String>>(),
+        };
+
         if let Some(readers) = self.external_resource_readers.as_mut() {
             readers.insert(key.into(), reader);
         } else {
-            let mut map = HashMap::new();
-            map.insert(key.into(), reader);
+            let map = HashMap::from([(key.into(), reader)]);
             self.external_resource_readers = Some(map);
         }
         self
@@ -154,13 +143,18 @@ impl EvaluatorOptions {
     pub fn external_module_reader(
         mut self,
         key: impl Into<String>,
-        reader: ExternalReader,
+        executable: impl Into<String>,
+        args: &[&str],
     ) -> Self {
+        let reader = ExternalReader {
+            executable: executable.into(),
+            arguments: args.into_iter().map(|s| s.to_string()).collect(),
+        };
+
         if let Some(readers) = self.external_module_readers.as_mut() {
             readers.insert(key.into(), reader);
         } else {
-            let mut map = HashMap::new();
-            map.insert(key.into(), reader);
+            let map = HashMap::from([(key.into(), reader)]);
             self.external_module_readers = Some(map);
         }
         self
@@ -314,7 +308,7 @@ pub fn start_pkl(pkl_debug: bool) -> Result<Child> {
     let mut command = Command::new("pkl");
 
     command
-        .arg("server") // Replace with actual command
+        .arg("server")
         .stdin(Stdio::piped())
         .stdout(Stdio::piped());
 
@@ -339,10 +333,8 @@ pub fn pkl_send_msg_one_way(
 
 pub fn decode_pkl_message(value: rmpv::Value) -> Option<PklServerMessage> {
     let decoded_array = value.as_array()?;
-
     let header = decoded_array.first()?.as_u64()?;
     let response = decoded_array.get(1)?.to_owned();
-
     Some(PklServerMessage { header, response })
 }
