+++ src/api/external_reader.rs
@@ -1,89 +1,220 @@
-use outgoing::PathElements;
+use std::io::Write;
 
-use super::evaluator::outgoing::{ClientModuleReader, ClientResourceReader};
+use crate::internal::msgapi::{
+    codes::{
+        CLOSE_EXTERNAL_PROCESS, INITIALIZE_MODULE_READER_REQUEST,
+        INITIALIZE_RESOURCE_READER_REQUEST, LIST_MODULES_REQUEST, LIST_RESOURCES_REQUEST,
+        READ_MODULE_REQUEST, READ_RESOURCE_REQUEST,
+    },
+    incoming::PklServerMessage,
+    outgoing::{
+        ClientModuleReader, ClientResourceReader, InitializeModuleReaderResponse,
+        InitializeResourceReaderResponse,
+    },
+    PklMessage,
+};
 
-pub mod incoming;
-pub mod outgoing;
-pub mod reader;
+use crate::{
+    api::{
+        evaluator::recv_msg,
+        reader::{
+            handle_list_modules, handle_list_resources, handle_read_module, handle_read_resource,
+        },
+    },
+    utils::macros::{_info, _warn},
+};
 
-// pub use incoming::ReadResource;
-
-pub trait PklReader {
-    fn read(&self);
-    fn list(&self);
+pub struct ExternalReaderRuntime {
+    resource_readers: Vec<Box<dyn PklResourceReader>>,
+    module_readers: Vec<Box<dyn PklModuleReader>>,
 }
 
-pub trait PklResourceReader {
-    /// Scheme returns the scheme part of the URL that this reader can read.
-    /// The value should be the URI scheme up to (not including) `:`
-    fn scheme(&self) -> &str;
-
-    /// Tells whether the path part of ths URI has a
-    /// [hier-part](https://datatracker.ietf.org/doc/html/rfc3986#section-3).
-    ///
-    /// An example of a hierarchical URI is `file:///path/to/my/file`, where
-    /// `/path/to/my/file` designates a nested path through the `/` character.
-    ///
-    /// An example of a non-hierarchical URI is `pkl.base`, where the `base` does not denote
-    /// any form of hierarchy.
-    fn has_hierarchical_uris(&self) -> bool {
-        false
-    }
-
-    /// Tells whether this reader supports globbing.
-    fn is_globbable(&self) -> bool {
-        false
-    }
-
-    fn read(&self, uri: &str) -> Result<Vec<u8>, Box<dyn std::error::Error>>;
-    fn list(&self, uri: &str) -> Result<Vec<PathElements>, Box<dyn std::error::Error>>;
-}
-
-pub trait PklModuleReader {
-    /// Scheme returns the scheme part of the URL that this reader can read.
-    /// The value should be the URI scheme up to (not including) `:`
-    fn scheme(&self) -> &str;
-
-    /// Tells whether the path part of ths URI has a
-    /// [hier-part](https://datatracker.ietf.org/doc/html/rfc3986#section-3).
-    ///
-    /// An example of a hierarchical URI is `file:///path/to/my/file`, where
-    /// `/path/to/my/file` designates a nested path through the `/` character.
-    ///
-    /// An example of a non-hierarchical URI is `pkl.base`, where the `base` does not denote
-    /// any form of hierarchy.
-    fn has_hierarchical_uris(&self) -> bool {
-        false
-    }
-
-    /// Tells whether this reader supports globbing.
-    fn is_globbable(&self) -> bool {
-        false
-    }
-
-    fn read(&self, uri: &str) -> Result<String, Box<dyn std::error::Error>>;
-    fn list(&self, uri: &str) -> Result<Vec<PathElements>, Box<dyn std::error::Error>>;
-
-    fn is_local(&self) -> bool;
-}
-
-impl From<&dyn PklModuleReader> for ClientModuleReader {
-    fn from(reader: &dyn PklModuleReader) -> Self {
-        ClientModuleReader {
-            scheme: reader.scheme().to_string(),
-            has_hierarchical_uris: reader.has_hierarchical_uris(),
-            is_globbable: reader.is_globbable(),
-            is_local: reader.is_local(),
+impl ExternalReaderRuntime {
+    pub fn new() -> Self {
+        Self {
+            resource_readers: Vec::new(),
+            module_readers: Vec::new(),
         }
     }
-}
 
-impl From<&dyn PklResourceReader> for ClientResourceReader {
-    fn from(reader: &dyn PklResourceReader) -> Self {
-        ClientResourceReader {
-            scheme: reader.scheme().to_string(),
-            has_hierarchical_uris: reader.has_hierarchical_uris(),
-            is_globbable: reader.is_globbable(),
+    /// Add a single, or tuple of resource readers to the client.
+    ///
+    /// Panics if any of the readers have the same scheme.
+    pub fn add_resource_readers<T: IntoResourceReaders>(&mut self, readers: T) -> &mut Self {
+        let readers = readers.into_readers();
+        self.resource_readers.extend(readers);
+
+        for (i, reader) in self.resource_readers.iter().enumerate() {
+            for other in &self.resource_readers[i + 1..] {
+                if reader.scheme() == other.scheme() {
+                    panic!(
+                        "Multiple resource readers sharing the same scheme: {}",
+                        reader.scheme()
+                    );
+                }
+            }
         }
+
+        self
+    }
+
+    pub fn add_module_readers<T: IntoModuleReaders>(&mut self, readers: T) -> &mut Self {
+        let readers = readers.into_readers();
+        self.module_readers.extend(readers);
+
+        for (i, reader) in self.module_readers.iter().enumerate() {
+            for other in &self.module_readers[i + 1..] {
+                if reader.scheme() == other.scheme() {
+                    panic!(
+                        "Multiple resource readers sharing the same scheme: {}",
+                        reader.scheme()
+                    );
+                }
+            }
+        }
+
+        self
+    }
+
+    fn handle_initalize_resource_reader<W: Write>(
+        &self,
+        pkl_msg: &PklServerMessage,
+        out: &mut W,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        debug_assert!(pkl_msg.header == INITIALIZE_RESOURCE_READER_REQUEST);
+
+        let map = pkl_msg.response.as_map().unwrap();
+        let request_id = map.get(0).unwrap().1.as_i64().unwrap();
+        let scheme = map.get(1).unwrap().1.as_str().unwrap();
+
+        // TODO: send error to pkl
+        let Some(reader) = self.resource_readers.iter().find(|r| r.scheme() == scheme) else {
+            _warn!("Incompatible scheme: {:?}", scheme);
+
+            let serialized = InitializeResourceReaderResponse {
+                request_id,
+                spec: None,
+            }
+            .encode_msg()?;
+
+            out.write_all(&serialized)?;
+            out.flush()?;
+
+            return Ok(());
+        };
+
+        let serialized = InitializeResourceReaderResponse {
+            request_id,
+            spec: Some(ClientResourceReader {
+                scheme: scheme.to_owned(),
+                has_hierarchical_uris: reader.has_hierarchical_uris(),
+                is_globbable: reader.is_globbable(),
+            }),
+        }
+        .encode_msg()?;
+
+        out.write_all(&serialized)?;
+        out.flush()?;
+
+        Ok(())
+    }
+
+    fn handle_initalize_module_reader<W: Write>(
+        &self,
+        pkl_msg: &PklServerMessage,
+        out: &mut W,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        debug_assert!(pkl_msg.header == INITIALIZE_MODULE_READER_REQUEST);
+
+        let map = pkl_msg.response.as_map().unwrap();
+        let request_id = map.get(0).unwrap().1.as_i64().unwrap();
+        let scheme = map.get(1).unwrap().1.as_str().unwrap();
+
+        // TODO: send error to pkl
+        let Some(reader) = self.module_readers.iter().find(|r| r.scheme() == scheme) else {
+            _warn!("Incompatible scheme: {:?}", scheme);
+
+            let serialized = InitializeModuleReaderResponse {
+                request_id,
+                spec: None,
+            }
+            .encode_msg()?;
+
+            out.write_all(&serialized)?;
+            out.flush()?;
+
+            return Ok(());
+        };
+
+        let serialized = InitializeModuleReaderResponse {
+            request_id,
+            spec: Some(ClientModuleReader {
+                scheme: scheme.to_owned(),
+                has_hierarchical_uris: reader.has_hierarchical_uris(),
+                is_globbable: reader.is_globbable(),
+                is_local: reader.is_local(),
+            }),
+        }
+        .encode_msg()?;
+
+        out.write_all(&serialized)?;
+        out.flush()?;
+
+        Ok(())
+    }
+
+    pub fn run(&self) -> Result<(), Box<dyn std::error::Error>> {
+        let mut stdin = std::io::stdin().lock();
+        let mut stdout = std::io::stdout().lock();
+
+        for _reader in self.resource_readers.iter() {
+            _info!("Registered resource reader: {:?}", _reader.scheme());
+        }
+
+        for _reader in self.module_readers.iter() {
+            _info!("Registered module reader: {:?}", _reader.scheme());
+        }
+
+        loop {
+            let Ok(pkl_msg) = recv_msg(&mut stdin) else {
+                _warn!("Failed to decode message");
+                break;
+            };
+
+            match pkl_msg.header {
+                INITIALIZE_RESOURCE_READER_REQUEST => {
+                    self.handle_initalize_resource_reader(&pkl_msg, &mut stdout)?;
+                }
+                INITIALIZE_MODULE_READER_REQUEST => {
+                    self.handle_initalize_module_reader(&pkl_msg, &mut stdout)?;
+                }
+                LIST_RESOURCES_REQUEST => {
+                    handle_list_resources(&self.resource_readers, &pkl_msg, &mut stdout)?;
+                }
+                LIST_MODULES_REQUEST => {
+                    handle_list_modules(&self.module_readers, &pkl_msg, &mut stdout)?;
+                }
+                READ_RESOURCE_REQUEST => {
+                    handle_read_resource(&self.resource_readers, &pkl_msg, &mut stdout)?;
+                }
+                READ_MODULE_REQUEST => {
+                    handle_read_module(&self.module_readers, &pkl_msg, &mut stdout)?;
+                }
+                CLOSE_EXTERNAL_PROCESS => {
+                    _info!("CLOSE_EXTERNAL_PROCESS received");
+                    break;
+                }
+                _ => {
+                    _warn!("unexpected message type: {:x}", pkl_msg.header);
+                }
+            }
+        }
+
+        Ok(())
     }
 }
+
+pub use crate::api::reader::{PklModuleReader, PklResourceReader};
+pub use crate::internal::msgapi::outgoing::PathElements;
+
+use super::reader::{IntoModuleReaders, IntoResourceReaders};
