+++ src/api/mod.rs
@@ -2,7 +2,6 @@ pub mod evaluator;
 pub use evaluator::Evaluator;
 pub(crate) mod decoder;
 pub mod external_reader;
-pub(crate) mod msgapi;
-pub(crate) mod reader;
+pub mod reader;
 
 pub use decoder::pkl_eval_module;
