+++ src/internal/msgapi/codes.rs
@@ -1,0 +1,20 @@
+pub const CREATE_EVALUATOR: u64 = 0x20;
+pub const CLOSE: u64 = 0x22;
+pub const EVALUATE_REQUEST: u64 = 0x23;
+
+pub const READ_RESOURCE_REQUEST: u64 = 0x26;
+pub const READ_RESOURCE_RESPONSE: u64 = 0x27;
+pub const READ_MODULE_REQUEST: u64 = 0x28;
+pub const READ_MODULE_RESPONSE: u64 = 0x29;
+
+pub const LIST_RESOURCES_REQUEST: u64 = 0x2a;
+pub const LIST_RESOURCES_RESPONSE: u64 = 0x2b;
+pub const LIST_MODULES_REQUEST: u64 = 0x2c;
+pub const LIST_MODULES_RESPONSE: u64 = 0x2d;
+
+pub const INITIALIZE_MODULE_READER_REQUEST: u64 = 0x2e;
+pub const INITIALIZE_MODULE_READER_RESPONSE: u64 = 0x2f;
+
+pub const INITIALIZE_RESOURCE_READER_REQUEST: u64 = 0x30;
+pub const INITIALIZE_RESOURCE_READER_RESPONSE: u64 = 0x31;
+pub const CLOSE_EXTERNAL_PROCESS: u64 = 0x32;
